requirements.txt
furl
requests
python-dateutil
tqdm
